Add parameter tests
