+++ tests/Parameter/CountTest.php
@@ -1,54 +1,40 @@
 <?php declare(strict_types=1);
 
-use lfischer\openWeatherMap\API;
-use lfischer\openWeatherMap\Endpoint\CurrentWeatherData;
-use lfischer\openWeatherMap\Endpoint\DailyForecastData;
-use lfischer\openWeatherMap\Endpoint\HourlyForecastData;
-use lfischer\openWeatherMap\RequestAdapter\Dump;
-use lfischer\openWeatherMap\RequestAdapter\RequestAdapterInterface;
+use lfischer\openWeatherMap\Parameter\CountTrait;
 use PHPUnit\Framework\TestCase;
-use Prophecy\Argument;
 
-final class ApiTest extends TestCase
+final class CountTest extends TestCase
 {
-    private const KEY = 'example-key';
+    use CountTrait;
 
-    /**
-     * @var API
-     */
-    private $api;
-    private $prophet;
-
-    /**
-     * This method is called before each test.
-     */
-    protected function setUp(): void
+    public function testSetNullCount()
     {
-        $this->api = new API(self::KEY);
-        $this->prophet = new \Prophecy\Prophet();
+        $this->setCount(null);
+
+        $this->assertEquals(['cnt' => null], $this->getCountParameter());
     }
 
-    public function testUrlFormatting()
+    public function testSetValidCount()
     {
-        $url = $this->api
-            ->setRequestAdapter(new Dump())
-            ->fetch('endpoint', ['a' => 1, 'b' => 2]);
+        $this->setCount(0);
+        $this->assertEquals(['cnt' => 1], $this->getCountParameter());
 
-        $this->assertEquals(API::URL . 'endpoint?appid=example-key&a=1&b=2', $url);
+        $this->setCount(1);
+        $this->assertEquals(['cnt' => 1], $this->getCountParameter());
+
+        $this->setCount(50);
+        $this->assertEquals(['cnt' => 50], $this->getCountParameter());
     }
 
-    public function testGetCurrentWeatherClient()
+    public function testSetInvalidFloat()
     {
-        $this->assertInstanceOf(CurrentWeatherData::class, $this->api->getCurrentWeatherClient());
+        $this->expectException(TypeError::class);
+        $this->setCount(1.1);
     }
 
-    public function testGetHourlyForecastClient()
+    public function testSetInvalidString()
     {
-        $this->assertInstanceOf(HourlyForecastData::class, $this->api->getHourlyForecastClient());
-    }
-
-    public function testGetDailyForecastClient()
-    {
-        $this->assertInstanceOf(DailyForecastData::class, $this->api->getDailyForecastClient());
+        $this->expectException(TypeError::class);
+        $this->setCount('1');
     }
 }
